extract printLogo into function, run animation while entropy loads
